requirements.txt
# Django фреймворк
Django==5.1.3

# Переменные окружения
python-dotenv==1.0.1

# Работа с изображениями
Pillow==10.4.0

# Админка
django-jazzmin==3.0.0

# Экспорт данных
django-import-export==4.1.0

# Текстовый редактор
django-ckeditor==6.7.1

# PostgreSQL драйвер (для продакшена)
psycopg2-binary==2.9.9

# Веб-сервер для продакшена
gunicorn==21.2.0

whitenoise==6.6.0
dj-database-url==2.1.0

# Дополнительные пакеты безопасности
# django-ratelimit==4.1.0  # Раскомментировать для rate limiting
# django-bleach==3.1.0     # Раскомментировать для санитизации HTML